Add currency to annual-wage as an attribute of an implicit relation (#38)
## What is the goal of this PR?

Add currency to annual-wage as an attribute of an implicit relation

## What are the changes implemented in this PR?

This is a simple extension of the query already present in the `EmploymentAgent`. Worth noting that this could be achieved by reasoning too, but we don't have a case when we do insertion onto an implicit relation like this, so better to keep it this way for now. Particularly since querying for `currency` in the query count test causes the test to time out, and therefore we need to benchmark this urgently!

resolves #35