Class hierarchy refactor (#76)
## What is the goal of this PR?

We refactor the structure of the simulation. There were code smells regarding unsafe castings and unnecessarily complicated abstraction of database connection/session/transaction behaviours (which couldn't be fully abstracted).

## What are the changes implemented in this PR?

The largest change is to convert the existing hierarchy of agents into a set of base interfaces that describe the behaviour of each agent type. The class hierarchy then follows the database-specific implementation. This way, the connection/session/transaction behaviour of each database can be encapsulated in the base classes `GraknAgent` and `Neo4jAgent`, without a need for a common abstraction of the two.