Refactor Agents to have Actions and DbOperations (#88)
## What is the goal of this PR?

While completing the comparison test there were bugs in the simulation queries that were difficult to uncover since testing was only taking place for the `insert` queries of each agent. To increase test granularity, and to remove some architectural code smells, a refactor was needed. `Agents` now have `Actions`. Actions are implemented on a per-db basis to determine the effects that they have and the queries they run. Transactional behaviour is abstracted as `DbOperations`, such that this is agnostic to transactions and when commits take place behaves consistently across different backends.

## What are the changes implemented in this PR?
This includes a sweeping refactor over almost the whole codebase. The responsibilities of each class should now be much clearer:
- Agents are now classes rather than interfaces with default methods.
- Actions have abstract boilerplate to determine the arguments they take and store them as protected variables.
- Actions have implementations specific to each DB, which is where the actual queries are located. This keeps the layer that changes per new db we add to be thin.
- Generics are used extensively to achieve the parameterisation of these classes for different DB types.
- Testing now checks the input and output of each action between the db implementations. This means that it's easy to see when an Action received different arguments between Grakn and Neo4j simulations and therefore that this Action's query isn't at fault, and a prior query is to blame for the discrepancy.
- We make use of the factory pattern to create Agents, and factories for Actions which have a subclass for each DB being tested.
- Simulation is now an abstract class and initialisation code now takes place in the `GraknSimulation` and `Neo4jSimulation`classes, which was previously disparate.