Emulate Reasoning with Property Graph Queries (#92)
## What is the goal of this PR?

To compare Agents between Grakn and property graphs we have to somehow account for their differences. Grakn has a reasoning engine, and so it can infer answers along a tree of reasoning.

To perform the equivalent in the property graphs we compare against, we have them make these trees of reasoning in their matching queries.

## What are the changes implemented in this PR?

This PR changes to the strategy described above, moving away from an ill-conceived strategy of materialising all facts in property graphs that will be inferred in the Grakn model.