Fix the issue with incorrect trace labeling when agents executed asynchronously (#116)
## What is the goal of this PR?

We fixed the issue when agents executed asynchronously were labeled incorrectly.

## What are the changes implemented in this PR?

Grabl tracing nests agents and their traces based on the thread those agents are executed on. That can lead to undefined behavior when threads are reused. To prevent that, we removed tracing from all the levels except the root one (1 trace per iteration for every agent).