Fix report collection for comparison test (#132)
## What is the goal of this PR?

Fix the collecting of reports for comparison testing.

## What are the changes implemented in this PR?

Agent reports are now stored by the simple name of the parent class. e.g. a report for `TypeDBPersonAgent` will be stored in the map of reports with the key `PersonAgent` so that the comparison test can easily find it based on the list of registered agents.