Parametrised database reinitialisation (#144)
## What is the goal of this PR?

Running multiple simulations on the same database without reinitialising
it allows simulations to be run incrementally, which is very useful for
exploring the way the data changes throughout the simulation for both
research and bug fixing. This PR makes that possible.

## What are the changes implemented in this PR?

The Config class has been modified to include a new `recreateDatabase`
parameter. The Context and Simulation classes have been updated
accordingly. In order to ensure that repeated runs of the same
simulation are deterministic, the Agent class has been modified so that
each spawned agent now receives its own `RandomSource`. The Simulation
class has also recieved a small modification to allow this, by ensuring
that the master random source is incremented regardless of whether the
database is recreated or not. Finally, the Simulation class has also
been modified to produce a slightly clearer output by including the
number of times each agent action is run.