Allowed simulations to specify multiple schema files (#146)
## What is the goal of this PR?

Allows the TypeQL used to define the schema for the TypeDB simulation to
be split over multiple files.

## What are the changes implemented in this PR?

Minor changes to the TypeDBSimulation class to replace the schema file
variable with a list of schema files and handle it accordingly. This is
a breaking change for any simulations depending on this repo.